pyproject.toml
[project]
name = "amazon-product-scraper-api"
version = "1.0.0"
description = "RESTful API for scraping Amazon product information by ASIN"
authors = [
    {name = "Amazon Product Scraper", email = "dev@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=2.3.3",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "fake-useragent>=1.4.0",
    "flask-limiter>=3.5.0",
    "redis>=5.0.1",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-flask>=1.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true